+++ src/components/Input/Input.story.js
@@ -23,6 +23,7 @@ stories
       <Input
         defaultValue={ text('defaultValue', 'some default value') }
         displayError={ boolean('displayError', false) }
+        hasSpellCheck={ boolean('hasSpellCheck', false) }
         id="input-01"
         isFilter={ boolean('isFilter', false) }
         isDropdown={ boolean('isDropdown', true) }
@@ -34,7 +35,6 @@ stories
         onChange={ action('on change') }
         onBlur={ action('on blur') }
         onKeyDown={ action('on key press') }
-        spellCheck={ boolean('spellCheck', false) }
         placeholder={ text('placeholder', 'just a placeholder') }
         isRequired={ boolean('isRequired', false) }
         type={ select('type', ['text', 'password', 'number', 'date'], 'text') }